+++ CHANGELOG.md
@@ -18,6 +18,7 @@ to [Semantic Versioning](http://semver.org/).
 - The `Person` API:
     - `chance.gender()`
     - `chance.age()`
+    - `chance.prefix()`
 
 ### Changed